+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Changelog
 
-All Notable changes to `command-validator` will be documented in this file.
+All notable changes to `command-validator` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+
+
+## 2.0.0 - 2019-04-30
+
+### Changed
+- Package dependencies
+- Code style
+- Integration tests
 
 
 ## 1.0.1 - 2016-09-29